Implementing menu subclass overrides
