+++ Fernanda/src/NotebookMenuModule.h
@@ -9,13 +9,16 @@
 
 #pragma once
 
+#include <QAction>
+#include <QHash>
+#include <QMenu>
 #include <QObject>
 
 #include "Coco/Debug.h"
 
 #include "Commander.h"
 #include "EventBus.h"
-#include "TempNewMenuModule.h"
+#include "MenuModule.h"
 
 namespace Fernanda {
 
@@ -39,18 +42,74 @@ public:
 protected:
     virtual void initializeWorkspaceActions_(Window* window) override
     {
-        //...
+        if (!window) return;
+        Actions_ actions{};
+
+        /// WIP
+        actions.fileImport = make(window, "", Tr::Menus::notebookFileImport());
+
+        /// WIP
+        actions.fileSave = make(window, "", Tr::Menus::notebookFileSave());
+
+        /// WIP
+        actions.fileSaveAs = make(window, "", Tr::Menus::notebookFileSaveAs());
+
+        /// WIP
+        actions.fileExport = make(window, "", Tr::Menus::notebookFileExport());
+
+        /// WIP
+        actions.fileOpenNotepad =
+            make(window, "", Tr::Menus::notebookFileOpenNotepad());
+
+        actions_[window] = actions;
+    }
+
+    [[nodiscard]]
+    virtual bool
+    addWorkspaceOpenActions_(QMenu* fileMenu, Window* window) override
+    {
+        if (!fileMenu || !window) return false;
+        auto& actions = actions_[window];
+
+        fileMenu->addAction(actions.fileImport);
+        return true;
+    }
+
+    [[nodiscard]]
+    virtual bool
+    addWorkspaceSaveActions_(QMenu* fileMenu, Window* window) override
+    {
+        if (!fileMenu || !window) return false;
+        auto& actions = actions_[window];
+
+        fileMenu->addAction(actions.fileSave);
+        fileMenu->addAction(actions.fileSaveAs);
+        fileMenu->addAction(actions.fileExport);
+        return true;
+    }
+
+    [[nodiscard]]
+    virtual bool
+    addWorkspaceMiscFileActions_(QMenu* fileMenu, Window* window) override
+    {
+        if (!fileMenu || !window) return false;
+        auto& actions = actions_[window];
+
+        fileMenu->addAction(actions.fileOpenNotepad);
+        return true;
     }
 
 private:
     struct Actions_
     {
-        QAction* fileSaveNotebook = nullptr;
-        QAction* fileSaveNotebookAs = nullptr;
         QAction* fileImport = nullptr;
+        QAction* fileSave = nullptr;
+        QAction* fileSaveAs = nullptr;
         QAction* fileExport = nullptr;
         QAction* fileOpenNotepad = nullptr;
     };
+
+    QHash<Window*, Actions_> actions_{};
 
     void initialize_()
     {
