+++ Fernanda/src/NotepadMenuModule.h
@@ -9,13 +9,16 @@
 
 #pragma once
 
+#include <QAction>
+#include <QHash>
+#include <QMenu>
 #include <QObject>
 
 #include "Coco/Debug.h"
 
 #include "Commander.h"
 #include "EventBus.h"
-#include "TempNewMenuModule.h"
+#include "MenuModule.h"
 
 namespace Fernanda {
 
@@ -39,7 +42,66 @@ public:
 protected:
     virtual void initializeWorkspaceActions_(Window* window) override
     {
-        //...
+        if (!window) return;
+        Actions_ actions{};
+
+        /// WIP
+        actions.fileOpen = make(window, "", Tr::Menus::notepadFileOpen());
+
+        /// WIP
+        actions.toggles.fileSave = make(
+            window,
+            Calls::Save,
+            Tr::Menus::notepadFileSave(),
+            Qt::CTRL | Qt::Key_S);
+
+        /// WIP
+        actions.toggles.fileSaveAs = make(
+            window,
+            Calls::SaveAs,
+            Tr::Menus::notepadFileSaveAs(),
+            Qt::CTRL | Qt::ALT | Qt::Key_S);
+
+        /// WIP
+        actions.toggles.fileSaveAllInWindow = make(
+            window,
+            Calls::SaveWindow,
+            Tr::Menus::notepadFileSaveAllInWindow());
+
+        /// WIP
+        actions.toggles.fileSaveAll = make(
+            window,
+            Calls::SaveAll,
+            Tr::Menus::notepadFileSaveAll(),
+            Qt::CTRL | Qt::SHIFT | Qt::Key_S);
+
+        actions_[window] = actions;
+        // setInitialToggleStates_(window);
+    }
+
+    [[nodiscard]]
+    virtual bool
+    addWorkspaceOpenActions_(QMenu* fileMenu, Window* window) override
+    {
+        if (!fileMenu || !window) return false;
+        auto& actions = actions_[window];
+
+        fileMenu->addAction(actions.fileOpen);
+        return true;
+    }
+
+    [[nodiscard]]
+    virtual bool
+    addWorkspaceSaveActions_(QMenu* fileMenu, Window* window) override
+    {
+        if (!fileMenu || !window) return false;
+        auto& actions = actions_[window];
+
+        fileMenu->addAction(actions.toggles.fileSave);
+        fileMenu->addAction(actions.toggles.fileSaveAs);
+        fileMenu->addAction(actions.toggles.fileSaveAllInWindow);
+        fileMenu->addAction(actions.toggles.fileSaveAll);
+        return true;
     }
 
 private:
@@ -55,6 +117,8 @@ private:
             QAction* fileSaveAll = nullptr;
         } toggles;
     };
+
+    QHash<Window*, Actions_> actions_{};
 
     void initialize_()
     {
